fix: Deterministic unit tests
The previous unit tests depended on several packages which included
themes, such as Adwaita; as well as programs, such as Firefox.

I modified the unit tests to not depend on anything which should help
distros that package Rust dependencies AND run unit tests (which I think
Debian does?).

The tests use a fake hierarchy of themes that we can control. All image
files are fake and zero bytes.